pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spec-doc-validator"
version = "0.1.0"
description = "Validate data files against spec documents using Claude AI"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "markdown-it-py>=3.0.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
specval = "spec_validator.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["spec_validator"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
